(refactor) Deleting REST API and using Modulo API to make function calls
It is not part of the scope of the project to expose a complete API on the internet/localhost, and therefore, the existence of the previous REST API has become unnecessary.

Now, using the RCP concept as an example, the API module has been recreated to expose Objects (which represent the Models in the Database), with their proper manipulation functions.

- Several changes were made to the codebase to match the new way of calling the API.