Sales: Cart handling functions created
The basic functions of: adding product to cart, editing product quantity in cart, removing product from cart, clearing cart, have all been implemented.

Now, the objective is to make the way of accessing the elements of the page more 'cleaner', in the dialogs that are opened by it.
For this, a new file was designated, state.py, to be responsible for saving the reference to all the elements that need to be shared on the page and its components.

This same sharing mode should be implemented later on the other pages.